- refactored to BigInteger

+++ src/main/kotlin/com/github/shmvanhouten/adventofcode2019/day22/CardShuffleInstructionParser.kt
@@ -1,6 +1,7 @@
 package com.github.shmvanhouten.adventofcode2019.day22
 
 import com.github.shmvanhouten.adventofcode2019.day22.Instructiontype.*
+import java.math.BigInteger
 
 fun parseInstructions(input: String): List<ShuffleInstruction> {
     return input.split('\n')
@@ -10,10 +11,10 @@ fun parseInstructions(input: String): List<ShuffleInstruction> {
 fun toShuffleInstruction(instruction: String): ShuffleInstruction {
     val words = instruction.split(' ')
     return when {
-        words[0] == "cut" -> ShuffleInstruction(CUT, words[1].toLong())
+        words[0] == "cut" -> ShuffleInstruction(CUT, BigInteger(words[1]))
         words[2] == "increment" -> ShuffleInstruction(
             DEAL_WITH_INCREMENT,
-            words[3].toLong()
+            BigInteger(words[3])
         )
         else -> ShuffleInstruction(DEAL_INTO_NEW_STACK)
     }
